I now want to start to implement parser support, so any serial telegram can be converted to a TCP socket and vice versa using the right parser
I will implement a zerro parser and an parser example.
The zerro parser will do nothing but memcopy of the
input buffer to the output buffer. Also it is called
using a function pointer, so by bending this pointer
the zeroparser can be "overloaded" with a parser
needed for the protocol to be converted.